requirements.txt
Flask==3.1.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
pymacaroons==0.13.0
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
Flask-OpenID==1.3.1
flask-cors==5.0.0
django-openid-auth==0.17
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.4.1
